+++ lib/plugins/aws/rollbackFunction.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const BbPromise = require('bluebird');
-const validate = require('../lib/validate');
+const validate = require('./lib/validate');
 const fetch = require('node-fetch');
 
 class AwsRollbackFunction {
