+++ lib/plugins/index.js
@@ -31,7 +31,7 @@ module.exports = [
   require('./aws/metrics.js'),
   require('./aws/remove/index.js'),
   require('./aws/rollback.js'),
-  require('./aws/rollbackFunction/index.js'),
+  require('./aws/rollbackFunction.js'),
   require('./aws/package/compile/layers.js'),
   require('./aws/package/compile/functions.js'),
   require('./aws/package/compile/events/schedule.js'),
